backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
azure-identity==1.18.0
azure-keyvault-secrets==4.9.0
requests==2.32.3
python-dotenv==1.0.1
